# Optimize SlidingWindowKeyedProcessFunction