--- a/java/feathub-udf/flink-udf/src/main/java/com/alibaba/feathub/flink/udf/aggregation/CountAggFunc.java
+++ b/java/feathub-udf/flink-udf/src/main/java/com/alibaba/feathub/flink/udf/aggregation/CountAggFunc.java
@@ -16,30 +16,46 @@
 
 package com.alibaba.feathub.flink.udf.aggregation;
 
+import org.apache.flink.api.common.typeinfo.TypeInformation;
+import org.apache.flink.api.common.typeinfo.Types;
 import org.apache.flink.table.api.DataTypes;
 import org.apache.flink.table.types.DataType;
 
 /** Aggregation function that counts the number of values. */
-public class CountAggFunc implements AggFunc<Object, Long> {
-    private long cnt = 0;
+public class CountAggFunc implements AggFunc<Object, Long, CountAggFunc.CountAccumulator> {
 
     @Override
-    public void reset() {
-        cnt = 0;
+    public void add(CountAccumulator accumulator, Object value, long timestamp) {
+        accumulator.cnt += 1;
     }
 
     @Override
-    public void aggregate(Object value, long timestamp) {
-        cnt += 1;
+    public void retract(CountAccumulator accumulator, Object value) {
+        accumulator.cnt -= 1;
     }
 
     @Override
-    public Long getResult() {
-        return cnt;
+    public Long getResult(CountAccumulator accumulator) {
+        return accumulator.cnt;
     }
 
     @Override
     public DataType getResultDatatype() {
         return DataTypes.BIGINT();
+    }
+
+    @Override
+    public CountAccumulator createAccumulator() {
+        return new CountAccumulator();
+    }
+
+    @Override
+    public TypeInformation<CountAccumulator> getAccumulatorTypeInformation() {
+        return Types.POJO(CountAccumulator.class);
+    }
+
+    /** Accumulator of {@link CountAccumulator}. */
+    public static class CountAccumulator {
+        public Long cnt = 0L;
     }
 }