## What is the purpose of the change

This PR update the SlidingWindowKeyedProcessFunction to use Flink states for internal data structures to avoid OOM in case of large key space.  It reduces the number of timers when the sliding window enables the skip_same_window_output option and enable_empty_window_output option. It also optimizes the SlidingWindowKeyedProcessFunction by introducing the retract method to the AggFunc.

We set up a benchmark to test the performance improvement of the optimization. The benchmark calculates the sum over sliding windows with both with skip_same_window_output and enable_empty_window_output enable. The window sizes are 2ms, 5ms, 10ms, 100ms, and 10s. The step size of the sliding window is 2ms. Before the optimization, the throughput of single parallelism is 58,305. After the optimization, the throughput of single parallelism goes up to 379,025, about 650% improvement.

The python test related to this PR is added in a follow-up PR #60 to avoid conflicting with the ongoing effect of refactoring the testing infra #59.

## Brief change log

- Introduce BlackHoleSink
- SlidingWindowKeyedProcessFunction uses Flink state and reducse timers
- Optimize SlidingWindowKeyedProcessFunction with retraction and iterator

## Does this pull request potentially affect one of the following parts:

  - Dependencies (does it add or upgrade a dependency): no
  - The public API: no

## Documentation

  - Does this pull request introduce a new feature? no
  - If yes, how is the feature documented? not applicable