@Sxnan 2023-01-10
@lindong28 I updated the PR and introduced an optimization to reduce the timer usage. Could you please take another look?